+++ supreme/verifier/src/commonTest/kotlin/examples/Readme-Verifier-min.kt
@@ -1,0 +1,4 @@
+package docs.config.minimal
+import at.asitplus.attestation.supreme.AttestationVerifier
+
+val verifier =/*(1)!*/ AttestationVerifier(makoto)